Compare pure ical values for recurrence changes
Otherwise, the value params are also compared - which can lead to
situations where one parameter may be a string, and in the other object
it’s an array of a string.

  {"tzid"=>"Australia/Melbourne"}
  {"tzid"=>["Australia/Melbourne"]}

The `value_ical` objects are timestamp strings - which is a more direct
(and in our case, more accurate) comparison.